change from modal to scene
