+++ frontend/src/scenes/data-warehouse/external/SourceModal.tsx
@@ -1,4 +1,4 @@
-import { LemonButton, LemonModal, LemonModalProps } from '@posthog/lemon-ui'
+import { LemonButton } from '@posthog/lemon-ui'
 import { useActions, useValues } from 'kea'
 import { FEATURE_FLAGS } from 'lib/constants'
 import { featureFlagLogic } from 'lib/logic/featureFlagLogic'
@@ -7,17 +7,15 @@ import postgresLogo from 'public/postgres-logo.svg'
 import stripeLogo from 'public/stripe-logo.svg'
 import zendeskLogo from 'public/zendesk-logo.png'
 
-import { DatawarehouseTableForm } from '../new_table/DataWarehouseTableForm'
+import { DatawarehouseTableForm } from '../new/DataWarehouseTableForm'
 import PostgresSchemaForm from './forms/PostgresSchemaForm'
 import SourceForm from './forms/SourceForm'
 import { SourceConfig } from './sourceModalLogic'
 import { sourceModalLogic } from './sourceModalLogic'
 
-interface SourceModalProps extends LemonModalProps {}
-
-export default function SourceModal(props: SourceModalProps): JSX.Element {
+export default function SourceModal(): JSX.Element {
     const { modalTitle, modalCaption } = useValues(sourceModalLogic)
-    const { onClear, onBack, onSubmit } = useActions(sourceModalLogic)
+    const { onBack, onSubmit } = useActions(sourceModalLogic)
     const { currentStep } = useValues(sourceModalLogic)
 
     const footer = (): JSX.Element | null => {
@@ -38,18 +36,14 @@ export default function SourceModal(props: SourceModalProps): JSX.Element {
     }
 
     return (
-        <LemonModal
-            {...props}
-            width={600}
-            onAfterClose={() => onClear()}
-            title={modalTitle}
-            description={modalCaption}
-            footer={footer()}
-        >
+        <>
+            <h3>{modalTitle}</h3>
+            <p>{modalCaption}</p>
             <FirstStep />
             <SecondStep />
             <ThirdStep />
-        </LemonModal>
+            {footer()}
+        </>
     )
 }
